requirements.txt
pyspark
delta-spark
faker
cassandra-driver
apache-airflow
psycopg2-binary